Add a Http proxy test
We got SSL handshake timeout error when we request https website via proxy, it's due to a Windows workaround (see libcat/libcat@af2c3257), but now the workaround does not work anymore, so we removed it to make sure that SSL handshake can work on Unix-like systems. And we add this test to prevent this problem occurring again.